requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Data Handling
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP Evaluation
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0
loguru>=0.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# Optional: For model training
accelerate>=0.24.0
deepspeed>=0.12.0
wandb>=0.16.0
